+++ CHANGELOG.md
@@ -6,13 +6,15 @@ Change Log
 
 ### Fixed
 
-- Error during app import (issue #21)
+- Error during app import (#21)
 - Fail to save file kept open during VS Code restart
+- Fail to activate the extension for new users with no settings (#87)
 
 ### Changes
 
 - Keep original indent of JSON files. Do not execute the auto format automatically.
 - Update dependencies `request-light`, `uuid`, `vscode-json-languageservice`, `vscode-languageclient`, `vscode-languageserver`, `tempy`
+- First environment configuration process is more friendly (#87)
 
 1.3.22
 ------