+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Change Log
 - Fail to save file kept open during VS Code restart
 - Fail to activate the extension for new users with no settings (#87)
 - Fail to display opensource app icons in treeview (#82)
+- Fail to load apps list if there is single app with multiple versions (#91)
 
 ### Changes