+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-1.3.24
+1.3.27
 ------
 
 ### Fixed
@@ -17,6 +17,7 @@ Change Log
 - Keep original indent of JSON files. Do not execute the auto format automatically.
 - Update dependencies `request-light`, `uuid`, `vscode-json-languageservice`, `vscode-languageclient`, `vscode-languageserver`, `tempy`
 - First environment configuration process is more friendly (#87)
+- Opensource apps rename to "Examples"
 
 1.3.22
 ------